PMM 2-12

+++ PMM/UT5 - Bases de Datos e Hilos/TiendasBuenas/app/src/main/java/com/ammd/tiendasbuenas/EventosListener.kt
@@ -1,5 +1,5 @@
 package com.ammd.tiendasbuenas
 
 interface EventosListener {
-    fun editar(id:Long)
+    fun editar(id:Int)
 }
